requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
cfgv==3.4.0
distlib==0.3.9
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
identify==2.6.2
idna==3.10
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
python-dotenv==1.0.1
python-telegram-bot==21.7
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
virtualenv==20.27.1
